WIP: Make quartz jobs not persistent.
