+++ rpki-validator/src/main/java/net/ripe/rpki/validator3/api/health/HealthController.java
@@ -36,7 +36,7 @@ import net.ripe.rpki.validator3.api.bgp.BgpPreviewService;
 import net.ripe.rpki.validator3.api.bgp.BgpRisDump;
 import net.ripe.rpki.validator3.api.trustanchors.TaStatus;
 import net.ripe.rpki.validator3.api.util.BuildInformation;
-import net.ripe.rpki.validator3.config.background.BackgroundJobs;
+import net.ripe.rpki.validator3.background.BackgroundJobs;
 import net.ripe.rpki.validator3.domain.TrustAnchors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
