+++ rpki-validator/src/main/java/net/ripe/rpki/validator3/background/RepositoryValidationJob.java
@@ -27,41 +27,47 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package net.ripe.rpki.validator3.adapter.jpa;
+package net.ripe.rpki.validator3.background;
 
 import lombok.Getter;
 import lombok.Setter;
-import net.ripe.rpki.validator3.domain.TrustAnchor;
-import net.ripe.rpki.validator3.domain.TrustAnchorValidationRun;
-import net.ripe.rpki.validator3.domain.validation.TrustAnchorValidationService;
-import org.quartz.*;
+import net.ripe.rpki.validator3.domain.RpkiRepository;
+import net.ripe.rpki.validator3.domain.RpkiRepositoryValidationRun;
+import net.ripe.rpki.validator3.domain.validation.RpkiRepositoryValidationService;
+import org.quartz.DisallowConcurrentExecution;
+import org.quartz.Job;
+import org.quartz.JobBuilder;
+import org.quartz.JobDetail;
+import org.quartz.JobExecutionContext;
+import org.quartz.JobExecutionException;
+import org.quartz.JobKey;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @DisallowConcurrentExecution
-public class QuartzTrustAnchorValidationJob implements Job {
+public class RepositoryValidationJob implements Job {
 
-    public static final String TRUST_ANCHOR_ID_KEY = "trustAnchorId";
+    public static final String RPKI_REPOSITORY_ID = "rpkiRepositoryId";
 
     @Autowired
-    private TrustAnchorValidationService validationService;
+    private RpkiRepositoryValidationService validationService;
 
     @Getter
     @Setter
-    private long trustAnchorId;
+    private long rpkiRepositoryId;
 
     @Override
     public void execute(JobExecutionContext context) throws JobExecutionException {
-        validationService.validate(trustAnchorId);
+        validationService.validateRpkiRepository(rpkiRepositoryId);
     }
 
-    static JobDetail buildJob(TrustAnchor trustAnchor) {
-        return JobBuilder.newJob(QuartzTrustAnchorValidationJob.class)
-            .withIdentity(getJobKey(trustAnchor))
-            .usingJobData(TRUST_ANCHOR_ID_KEY, trustAnchor.getId())
+    static JobDetail buildJob(RpkiRepository rpkiRepository) {
+        return JobBuilder.newJob(RepositoryValidationJob.class)
+            .withIdentity(getJobKey(rpkiRepository))
+            .usingJobData(RPKI_REPOSITORY_ID, rpkiRepository.getId())
             .build();
     }
 
-    static JobKey getJobKey(TrustAnchor trustAnchor) {
-        return new JobKey(String.format("%s#%s", TrustAnchorValidationRun.TYPE, trustAnchor.getName()));
+    static JobKey getJobKey(RpkiRepository rpkiRepository) {
+        return new JobKey(String.format("%s#%s#%d", RpkiRepositoryValidationRun.TYPE, rpkiRepository.getRrdpNotifyUri(), rpkiRepository.getId()));
     }
 }
