+++ rpki-validator/src/main/resources/application.properties
@@ -42,6 +42,8 @@ spring.datasource.name=rpki
 spring.datasource.username=rpki
 spring.datasource.password=rpki
 
+org.quartz.jobStore.class=org.quartz.simpl.RAMJobStore
+
 spring.mvc.message-codes-resolver.format=POSTFIX_ERROR_CODE
 
 spring.jpa.hibernate.ddl-auto=validate
